requirements.txt
langchain
langchain_community
langchain_huggingface
sentence-transformers
huggingface_hub
python-dotenv
chromadb
pypdf
streamlit
